Render team screen. Fragment transitions.

+++ app/src/main/java/app/nba/app/data/net/dto/TeamDto.kt
@@ -3,7 +3,7 @@ package app.nba.app.data.net.dto
 import com.google.gson.annotations.SerializedName
 
 data class TeamDto(
-    @SerializedName("id") val id: Int? = 0,
+    @SerializedName("id") val id: Long? = 0,
     @SerializedName("abbreviation") val abbreviation: String? = "",
     @SerializedName("city") val city: String? = "",
     @SerializedName("conference") val conference: String? = "",
